requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
